Improved StreamWishExtractor and Added FileMoon Extractor Separately and Improved also Filsim
